% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Data preparation for CNN1D SWRC modelling}
\description{
Functions to prepare soil data for input to the CNN1D
monotone-integral model, including 3-D sequence array construction and
observation matrix creation.
}
