% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_bd_units}
\alias{fix_bd_units}
\title{Detect and correct bulk-density units}
\usage{
fix_bd_units(bd_raw)
}
\arguments{
\item{bd_raw}{Numeric vector of raw bulk-density values.}
}
\value{
Numeric vector in g/cm3.
}
\description{
If the median raw value is > 10 it is assumed to be in kg/m3 and is
divided by 100 to convert to g/cm3.
}
\examples{
fix_bd_units(c(1.2, 1.45, 1.3))   # already g/cm3
fix_bd_units(c(120, 145, 130))    # kg/m3 -> g/cm3

}
