% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{head_from_pf}
\alias{head_from_pf}
\title{Convert pF to matric head (cm)}
\usage{
head_from_pf(pf)
}
\arguments{
\item{pf}{Numeric vector of pF values.}
}
\value{
Numeric vector of matric head values in cm.
}
\description{
Convert pF to matric head (cm)
}
\examples{
head_from_pf(c(0, 1, 2, 3, 4.2))

}
