% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pf_from_head}
\alias{pf_from_head}
\title{Convert matric head (cm) to pF}
\usage{
pf_from_head(h_cm)
}
\arguments{
\item{h_cm}{Numeric vector of matric head values in cm (positive).}
}
\value{
Numeric vector of pF values (\eqn{\log_{10}(h)}).
}
\description{
Convert matric head (cm) to pF
}
\examples{
pf_from_head(c(1, 10, 100, 1000, 15850))

}
