% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pf_normalize}
\alias{pf_normalize}
\title{Normalise pF values to [0, 1]}
\usage{
pf_normalize(pf, pf_left = -2, pf_right = 7.6)
}
\arguments{
\item{pf}{Numeric vector of pF values.}

\item{pf_left}{Left boundary of the pF domain (default \code{-2}).}

\item{pf_right}{Right boundary of the pF domain (default \code{7.6}).}
}
\value{
Numeric vector in \verb{[0, 1]}.
}
\description{
Maps the pF domain \verb{[pf_left, pf_right]} linearly to \verb{[0, 1]}.
Values outside the domain are clipped.
}
\examples{
pf_normalize(c(-2, 0, 4, 7.6))

}
