% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{soilFlux-package}
\alias{soilFlux}
\alias{soilFlux-package}
\title{soilFlux: Physics-Informed Neural Networks for Soil Water Retention Curves}
\description{
The \code{soilFlux} package implements a physics-informed 1-D convolutional
neural network (CNN1D-PINN) for estimating the complete soil water
retention curve (SWRC) as a continuous function of matric potential,
from soil texture, organic carbon, bulk density, and depth.
}
\details{
\subsection{Main functions}{\tabular{ll}{
   Function \tab Purpose \cr
   \code{\link[=prepare_swrc_data]{prepare_swrc_data()}} \tab Standardise raw soil data \cr
   \code{\link[=fit_swrc]{fit_swrc()}} \tab Train the CNN1D-PINN model \cr
   \code{\link[=predict_swrc]{predict_swrc()}} \tab Predict theta at given pF values \cr
   \code{\link[=predict_swrc_dense]{predict_swrc_dense()}} \tab Predict full SWRC curves \cr
   \code{\link[=swrc_metrics]{swrc_metrics()}} \tab Evaluate model performance \cr
   \code{\link[=plot_swrc]{plot_swrc()}} \tab Plot retention curves \cr
   \code{\link[=plot_pred_obs]{plot_pred_obs()}} \tab Predicted vs. observed plot \cr
   \code{\link[=save_swrc_model]{save_swrc_model()}} / \code{\link[=load_swrc_model]{load_swrc_model()}} \tab Persist the model \cr
   \code{\link[=classify_texture]{classify_texture()}} \tab USDA texture classification \cr
}

}

\subsection{References}{

Norouzi, A. M., et al. (2025). "Physics-Informed Neural Networks for
Estimating a Continuous Form of the Soil Water Retention Curve."
\emph{Water Resources Research}. \doi{10.1029/2024WR038149}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HugoMachadoRodrigues/soilFlux}
  \item \url{https://hugomachadorodrigues.github.io/soilFlux/}
  \item Report bugs at \url{https://github.com/HugoMachadoRodrigues/soilFlux/issues}
}

}
\author{
\strong{Maintainer}: Hugo Rodrigues \email{rodrigues.machado.hugo@gmail.com} (\href{https://orcid.org/0000-0002-8070-8126}{ORCID})

}
\keyword{internal}
