% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{swrc_model_exists}
\alias{swrc_model_exists}
\title{Check whether a model directory contains a valid saved model}
\usage{
swrc_model_exists(dir = "./models/swrc", name = "swrc_model")
}
\arguments{
\item{dir}{Directory path.}

\item{name}{Model stem name.}
}
\value{
Logical scalar.
}
\description{
Check whether a model directory contains a valid saved model
}
