% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Training the CNN1D SWRC model}
\description{
High-level and low-level functions for training the
physics-informed CNN1D model, including the eager-mode train step
and the main fitting loop with early stopping.
}
