% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{print.swrc_fit}
\alias{print.swrc_fit}
\title{Print method for swrc_fit}
\usage{
\method{print}{swrc_fit}(x, ...)
}
\arguments{
\item{x}{An \code{swrc_fit} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x} (called for its side effect of printing a
summary of the fitted model to the console).
}
\description{
Print method for swrc_fit
}
