% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{summary.swrc_fit}
\alias{summary.swrc_fit}
\title{Summary method for swrc_fit}
\usage{
\method{summary}{swrc_fit}(object, ...)
}
\arguments{
\item{object}{An \code{swrc_fit} object.}

\item{...}{Ignored.}
}
\description{
Summary method for swrc_fit
}
